+++ packages/docs/src/index.ts
@@ -2,36 +2,33 @@ import ATTRIBUTES_DATA from '../attributes.json';
 import { ATTRIBUTES } from './constants';
 import { initCopyExampleButtons } from './copy-examples';
 import { initCopyScriptButton } from './copy-script';
-import { initExampleDropdowns } from './example-dropdowns';
 import { initTitles } from './titles';
 
 import type { AttributeExample } from './types';
-
-// Constants destructuring
-const {
-  attribute: { key: attributeKey },
-} = ATTRIBUTES;
 
 /**
  * Init
  */
 window.Webflow ||= [];
 window.Webflow.push(async () => {
-  initTitles();
-  initExampleDropdowns();
-
-  const currentAttributeKey = document.body.getAttribute(attributeKey);
+  const currentAttributeKey = document.body.getAttribute(ATTRIBUTES.attribute.key);
   if (!currentAttributeKey) return;
 
   const attributeData = ATTRIBUTES_DATA.find(({ key }) => key === currentAttributeKey);
   if (!attributeData) return;
 
+  initTitles();
+
   const { baseSrc, examplesSrc } = attributeData;
 
-  const response = await fetch(`${baseSrc}/${examplesSrc}`);
+  try {
+    const response = await fetch(`${baseSrc}/${examplesSrc}`);
 
-  const examples: AttributeExample[] = await response.json();
+    const examples: AttributeExample[] = await response.json();
 
-  initCopyExampleButtons(examples);
-  initCopyScriptButton(attributeData);
+    initCopyExampleButtons(examples);
+    initCopyScriptButton(attributeData);
+  } catch (error) {
+    return;
+  }
 });