+++ packages/docs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # `docs` Changelog
 
+## [v1.3.0] 14th October 2021
+
+- Removed unused code bits.
+- Added `try`/`catch` check on the attributes data fetching.
+- Made sure the script can be placed in the global Custom Code section instead of each Attribute's page.
+- Added example copy validity check.
+
 ## [v1.2.1] 25th September 2021
 
 - Created `CHANGELOG.md`.